+++ Marlin/src/gcode/bedlevel/abl/M421.cpp
@@ -42,7 +42,7 @@ extern uint32_t  GRID_MAX_POINTS_Y;
  *   M421 I<xindex> J<yindex> Q<offset>
  */
 void GcodeSuite::M421() {
-  int8_t ix = parser.intval('I', -1), iy = parser.intval('J', -1);
+  uint32_t ix = parser.intval('I', -1), iy = parser.intval('J', -1);
   const bool hasI = ix >= 0,
              hasJ = iy >= 0,
              hasZ = parser.seen('Z'),
